backend_new/tasks/update_tasks.py
"""Celery Tasks - Sources Update

Tareas para actualizar normativas, catálogos y bases de conocimiento
"""

from core.celery_app import celery_app
import logging
import yaml
from pathlib import Path

logger = logging.getLogger(__name__)

@celery_app.task(name="tasks.update_tasks.update_normativas")
def update_normativas():
    """Actualiza normativas oficiales (GTC 45, Decreto 1072, etc.)"""
    try:
        logger.info("Iniciando actualización de normativas...")
        
        # TODO: Implementar lógica de actualización
        # 1. Leer sources_config.yaml
        # 2. Descargar PDFs/scraping de páginas oficiales
        # 3. Guardar en esquema Silver
        
        logger.info("Normativas actualizadas correctamente")
        return {"status": "success", "updated": 0}
        
    except Exception as e:
        logger.error(f"Error actualizando normativas: {str(e)}")
        raise

@celery_app.task(name="tasks.update_tasks.update_catalogos")
def update_catalogos():
    """Actualiza catálogos de peligros y controles"""
    try:
        logger.info("Iniciando actualización de catálogos...")
        
        # TODO: Implementar lógica de actualización
        # 1. Leer archivos JSON locales
        # 2. Consultar APIs externas si están configuradas
        # 3. Actualizar tablas en esquema Silver
        
        logger.info("Catálogos actualizados correctamente")
        return {"status": "success", "updated": 0}
        
    except Exception as e:
        logger.error(f"Error actualizando catálogos: {str(e)}")
        raise

@celery_app.task(name="tasks.update_tasks.learn_from_matrices")
def learn_from_matrices():
    """Aprende patrones de matrices anteriores para mejorar sugerencias"""
    try:
        logger.info("Iniciando aprendizaje de matrices previas...")
        
        # TODO: Implementar lógica de aprendizaje
        # 1. Consultar matrices en esquema Gold
        # 2. Extraer patrones comunes (peligros frecuentes por sector, controles efectivos)
        # 3. Actualizar base de conocimiento en Silver
        
        logger.info("Aprendizaje completado")
        return {"status": "success", "patterns_learned": 0}
        
    except Exception as e:
        logger.error(f"Error en aprendizaje: {str(e)}")
        raise

@celery_app.task(name="tasks.update_tasks.update_all_sources")
def update_all_sources():
    """Ejecuta actualización completa de todas las fuentes"""
    try:
        logger.info("Iniciando actualización completa de fuentes...")
        
        results = {
            "normativas": update_normativas(),
            "catalogos": update_catalogos(),
            "learning": learn_from_matrices()
        }
        
        logger.info("Actualización completa finalizada")
        return {"status": "success", "results": results}
        
    except Exception as e:
        logger.error(f"Error en actualización completa: {str(e)}")
        raise