backend_new/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
pydantic==2.7.0
pydantic-settings==2.2.1

# Database
SQLAlchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9

# Async & Task Queue
celery==5.3.6
redis==5.0.3

# LangChain & LangGraph
langchain==0.1.20
langgraph==0.0.55

# AI/ML
emergentintegrations==0.1.0

# Document Processing
PyMuPDF==1.24.2
python-docx==1.1.0
openpyxl==3.1.2
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
typer==0.12.3

# Testing
pytest==8.1.1
httpx==0.27.0
